refactor: split metalsLspService to main and fallback parts
