+++ tests/unit/src/main/scala/tests/TestingClient.scala
@@ -322,7 +322,10 @@ class TestingClient(workspace: AbsolutePath, val buffers: Buffers)
       val buildTools = BuildTools.default().allAvailable
       buildTools.exists(newBuildTool =>
         buildTools.exists(oldBuildTool =>
-          NewBuildToolDetected.params(newBuildTool, oldBuildTool) == params
+          NewBuildToolDetected.params(
+            newBuildTool.executableName,
+            oldBuildTool.executableName,
+          ) == params
         )
       )
     }
